Conflicts:
	clients/Denominations.md
	clients/coffeescript/sputnik.coffee
	clients/python/load_tester.py
	clients/python/market_maker.py
	clients/python/random_trader.py
	clients/python/sputnik.py
	server/sputnik/accountant.py
	server/sputnik/admin_templates/base.html
	server/sputnik/administrator.py
	server/sputnik/util.py
	server/sputnik/webserver.py
	server/sputnik/webserver/plugins/db/postgres.py
	server/sputnik/webserver/plugins/receiver/administrator.py
	testing/test_administrator.py

+++ server/sputnik/util.py
@@ -16,6 +16,7 @@ import models
 from zmq_util import ComponentExport
 from sqlalchemy.orm.session import Session
 import hashlib
+from decimal import Decimal
 
 #
 # This doesn't work properly
@@ -90,17 +91,17 @@ def price_to_wire(contract, price):
 
 def price_from_wire(contract, price):
     if contract.contract_type == "cash_pair":
-        return float(price) / (contract.denominated_contract.denominator * contract.denominator)
+        return Decimal(price) / (contract.denominated_contract.denominator * contract.denominator)
     else:
-        return float(price) / contract.denominator
+        return Decimal(price) / contract.denominator
 
 def quantity_from_wire(contract, quantity):
     if contract.contract_type == "prediction" or contract.contract_type == "futures":
         return quantity
     elif contract.contract_type == "cash":
-        return float(quantity) / contract.denominator
+        return Decimal(quantity) / contract.denominator
     else:
-        return float(quantity) / contract.payout_contract.denominator
+        return Decimal(quantity) / contract.payout_contract.denominator
 
 def quantity_to_wire(contract, quantity):
     if contract.contract_type == "prediction" or contract.contract_type == "futures":
@@ -166,10 +167,10 @@ def timestamp_to_dt(timestamp):
 
 def get_cash_spent(contract, price, quantity):
     if contract.contract_type == "futures" or contract.contract_type == "prediction":
-        cash_float = float(quantity * price * contract.lot_size) / contract.denominator
+        cash_float = Decimal(quantity * price * contract.lot_size) / contract.denominator
     else:
         payout_contract = contract.payout_contract
-        cash_float = float(quantity * price) / (contract.denominator * payout_contract.denominator)
+        cash_float = Decimal(quantity * price) / (contract.denominator * payout_contract.denominator)
 
     cash_int = int(cash_float)
     if cash_float != cash_int:
@@ -216,8 +217,8 @@ def get_deposit_fees(user, contract, deposit_amount, trial_period=False):
     if trial_period:
         return {}
 
-    base_fee = contract.deposit_base_fee + float(deposit_amount * contract.deposit_bps_fee) / 10000
-    user_factor = float(user.fees.deposit_factor) / 100
+    base_fee = contract.deposit_base_fee + Decimal(deposit_amount * contract.deposit_bps_fee) / 10000
+    user_factor = Decimal(user.fees.deposit_factor) / 100
     final_fee = int(round(base_fee * user_factor))
 
     return {contract.ticker: final_fee}
@@ -226,8 +227,8 @@ def get_withdraw_fees(user, contract, withdraw_amount, trial_period=False):
     if trial_period:
         return {}
 
-    base_fee = contract.withdraw_base_fee + float(withdraw_amount * contract.withdraw_bps_fee) / 10000
-    user_factor = float(user.fees.withdraw_factor) / 100
+    base_fee = contract.withdraw_base_fee + Decimal(withdraw_amount * contract.withdraw_bps_fee) / 10000
+    user_factor = Decimal(user.fees.withdraw_factor) / 100
     final_fee = int(round(base_fee * user_factor))
 
     return {contract.ticker: final_fee}
